+++ R/assign-age-groups.R
@@ -50,7 +50,11 @@ assign_age_groups <- function(
   check_if_contact_survey(survey)
   check_age_limits_increasing(age_limits)
   estimated_participant_age <- rlang::arg_match(estimated_participant_age)
-  estimated_contact_age <- rlang::arg_match(estimated_contact_age)
+  if (is.data.frame(estimated_contact_age)) {
+    estimated_contact_age <- validate_age_distribution(estimated_contact_age)
+  } else {
+    estimated_contact_age <- rlang::arg_match(estimated_contact_age)
+  }
   missing_participant_age <- rlang::arg_match(missing_participant_age)
   missing_contact_age <- rlang::arg_match(missing_contact_age)